// src/types/api.d.ts
export interface components {
    schemas: {
        CreateProductDto: {
            sku: string;
            displayName: string;
            description?: string;
            price: number;
            isOnSale?: boolean;
        };
        Product: {
            uuid: string;
            sku: string;
            displayName: string;
            description?: string;
            /** @default 10.00 */
            price: string;
            /** @default false */
            isOnSale: boolean;
            /**
             * Format: date-time
             * @default 2025-12-16T10:27:10.218Z
             */
            createdAt: string;
            /**
             * Format: date-time
             * @default 2025-12-16T10:27:10.218Z
             */
            updatedAt: string;
            /** Format: date-time */
            deletedAt?: string;
            /** @default 0 */
            stockAmount: number;
        };
        UpdateProductDto: {
            sku?: string;
            displayName?: string;
            description?: string;
            price?: number;
            isOnSale?: boolean;
        };
        AddStockDto: {
            /** Format: uuid */
            productUuid: string;
            quantity: number;
            notes?: string;
        };
        StockTransaction: {
            uuid: string;
            product: components["schemas"]["Product"];
            quantity: number;
            /** @enum {string} */
            type: "INCOMING" | "OUTGOING_ORDER" | "ADJUSTMENT";
            referenceId?: string;
            notes?: string;
            /**
             * Format: date-time
             * @default 2025-12-16T10:27:10.221Z
             */
            createdAt: string;
        };
        AdjustStockDto: {
            /** Format: uuid */
            productUuid: string;
            quantity: number;
            notes?: string;
        };
        CreateOrderItemDto: {
            /** Format: uuid */
            productUuid: string;
            quantity: number;
        };
        CreateOrderDto: {
            customerName?: string;
            /** Format: email */
            customerEmail: string;
            items: components["schemas"]["CreateOrderItemDto"][];
        };
        UpdateOrderStatusDto: {
            /** @enum {string} */
            status: "PENDING" | "CONFIRMED" | "CANCELLED" | "FULFILLED";
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}
export type CreateProductDto = components['schemas']['CreateProductDto'];
export type Product = components['schemas']['Product'];
export type UpdateProductDto = components['schemas']['UpdateProductDto'];
export type AddStockDto = components['schemas']['AddStockDto'];
export type StockTransaction = components['schemas']['StockTransaction'];
export type AdjustStockDto = components['schemas']['AdjustStockDto'];
export type CreateOrderItemDto = components['schemas']['CreateOrderItemDto'];
export type CreateOrderDto = components['schemas']['CreateOrderDto'];
export type UpdateOrderStatusDto = components['schemas']['UpdateOrderStatusDto'];